提现部分

+++ mall-spring-boot/src/main/java/edu/nju/mall/entity/WithdrawalCondition.java
@@ -28,5 +28,5 @@ public class WithdrawalCondition {
 
     private int member;
 
-    private double cash;
+    private Long cash;
 }
